+++ src/hsmodules/ManagedCare/Referral.js
@@ -1,31 +1,31 @@
 /* eslint-disable */
-import React, { useState, useContext, useEffect, useRef } from 'react';
-import { Route, useNavigate, Link, NavLink } from 'react-router-dom';
-import client from '../../feathers';
-import { DebounceInput } from 'react-debounce-input';
-import { useForm } from 'react-hook-form';
-import { UserContext, ObjectContext } from '../../context';
-import { toast } from 'bulma-toast';
-import { formatDistanceToNowStrict, format, subDays, addDays } from 'date-fns';
-import DatePicker from 'react-datepicker';
-import LocationSearch from '../helpers/LocationSearch';
-import EmployeeSearch from '../helpers/EmployeeSearch';
-import BillServiceCreate from '../Finance/BillServiceCreate';
-import 'react-datepicker/dist/react-datepicker.css';
-import AddCircleOutline from '@mui/icons-material/AddCircleOutline';
-import { GridBox } from '../app/styles';
-import { PageWrapper } from '../../ui/styled/styles';
-import { TableMenu } from '../../ui/styled/global';
-import FilterMenu from '../../components/utilities/FilterMenu';
-import Button from '../../components/buttons/Button';
-import CustomTable from '../../components/customtable';
-import Switch from '../../components/switch';
-import { BsFillGridFill, BsList, BsCheckCircle } from 'react-icons/bs';
-import { FaRegTimesCircle } from 'react-icons/fa';
-import { BiPencil } from 'react-icons/bi';
-import CalendarGrid from '../../components/calender';
-import ModalBox from '../../components/modal';
-import ModalHeader from '../Appointment/ui-components/Heading/modalHeader';
+import React, { useState, useContext, useEffect, useRef } from "react";
+import { Route, useNavigate, Link, NavLink } from "react-router-dom";
+import client from "../../feathers";
+import { DebounceInput } from "react-debounce-input";
+import { useForm } from "react-hook-form";
+import { UserContext, ObjectContext } from "../../context";
+import { toast } from "bulma-toast";
+import { formatDistanceToNowStrict, format, subDays, addDays } from "date-fns";
+import DatePicker from "react-datepicker";
+import LocationSearch from "../helpers/LocationSearch";
+import EmployeeSearch from "../helpers/EmployeeSearch";
+import BillServiceCreate from "../Finance/BillServiceCreate";
+import "react-datepicker/dist/react-datepicker.css";
+import AddCircleOutline from "@mui/icons-material/AddCircleOutline";
+import { GridBox } from "../app/styles";
+import { PageWrapper } from "../../ui/styled/styles";
+import { TableMenu } from "../../ui/styled/global";
+import FilterMenu from "../../components/utilities/FilterMenu";
+import Button from "../../components/buttons/Button";
+import CustomTable from "../../components/customtable";
+import Switch from "../../components/switch";
+import { BsFillGridFill, BsList, BsCheckCircle } from "react-icons/bs";
+import { FaRegTimesCircle } from "react-icons/fa";
+import { BiPencil } from "react-icons/bi";
+import CalendarGrid from "../../components/calender";
+import ModalBox from "../../components/modal";
+import ModalHeader from "../Appointment/ui-components/Heading/modalHeader";
 import {
   Radio,
   Grid,
@@ -34,43 +34,59 @@ import {
   Box,
   Badge,
   Drawer,
-} from '@mui/material';
-import DebouncedInput from '../Appointment/ui-components/inputs/DebouncedInput';
-import { McText } from './text';
-import Input from '../../components/inputs/basic/Input/index';
-import ToggleButton from '../../components/toggleButton';
-import RadioButton from '../../components/inputs/basic/Radio';
-import BasicDatePicker from '../../components/inputs/Date';
-import BasicDateTimePicker from '../../components/inputs/DateTime';
-import CustomSelect from '../../components/inputs/basic/Select';
-import Textarea from '../../components/inputs/basic/Textarea';
-import { MdCancel, MdAddCircle } from 'react-icons/md';
-import PatientProfile from '../Client/PatientProfile';
-import ManagedCareLeft from './components/manageCareRight';
-import GlobalCustomButton from '../../components/buttons/CustomButton';
-import { FormsHeaderText } from '../../components/texts';
-import ChatInterface from '../../components/chat/ChatInterface';
-import CRMTasks from '../CRM/Tasks';
-import AutoCompleteBox from '../../components/inputs/AutoComplete';
-import { FacilitySearch } from '../helpers/FacilitySearch';
+} from "@mui/material";
+import DebouncedInput from "../Appointment/ui-components/inputs/DebouncedInput";
+import { McText } from "./text";
+import Input from "../../components/inputs/basic/Input/index";
+import ToggleButton from "../../components/toggleButton";
+import RadioButton from "../../components/inputs/basic/Radio";
+import BasicDatePicker from "../../components/inputs/Date";
+import BasicDateTimePicker from "../../components/inputs/DateTime";
+import CustomSelect from "../../components/inputs/basic/Select";
+import Textarea from "../../components/inputs/basic/Textarea";
+import { MdCancel, MdAddCircle } from "react-icons/md";
+import PatientProfile from "../Client/PatientProfile";
+import ManagedCareLeft from "./components/manageCareRight";
+import GlobalCustomButton from "../../components/buttons/CustomButton";
+import { FormsHeaderText } from "../../components/texts";
+import ChatInterface from "../../components/chat/ChatInterface";
+import CRMTasks from "../CRM/Tasks";
+import AutoCompleteBox from "../../components/inputs/AutoComplete";
+import { FacilitySearch } from "../helpers/FacilitySearch";
+import { ReferralList } from "./components/referral/ReferralList";
+import {ReferralCreate} from './components/referral/CreateReferral'
 
 // eslint-disable-next-line
 const searchfacility = {};
 
-export default function GeneralAppointments() {
-  const { state } = useContext(ObjectContext); //,setState
-  // eslint-disable-next-line
-  const [selectedClient, setSelectedClient] = useState();
-  const [selectedAppointment, setSelectedAppointment] = useState();
-  //const [showState,setShowState]=useState() //create|modify|detail
+export default function Referral() {
+  const { state, setState } = useContext(ObjectContext);
+  const [view, setView] = useState("list");
   const [showModal, setShowModal] = useState(0);
 
+  const handleGoBack = () => {
+    setView("list");
+    setState((prev) => ({
+      ...prev,
+      ClientModule: {
+        ...prev.ClientModule,
+        selectedClient: {},
+      },
+    }));
+  };
+
   return (
-    <section className="section remPadTop">
-      {showModal === 0 && (
-        <ReferralList showModal={showModal} setShowModal={setShowModal} />
+    <Box>
+      {view === "list" && (
+        <ReferralList
+          showCreate={() => setView("create")}
+          showDetail={() => setView("detail")}
+        />
       )}
-      {showModal === 1 && (
+
+      {view === "create" && <ReferralCreate  handleGoBack={handleGoBack}/>}
+
+      {/* {showModal === 1 && (
         <Grid container spacing={2}>
           <Grid item xs={3}>
             <PatientProfile />
@@ -79,7 +95,7 @@ export default function GeneralAppointments() {
             <ReferralCreate showModal={showModal} setShowModal={setShowModal} />
           </Grid>
         </Grid>
-      )}
+      )} */}
       {showModal === 2 && (
         <Grid container spacing={2}>
           <Grid item xs={3}>
@@ -90,1105 +106,1106 @@ export default function GeneralAppointments() {
           </Grid>
         </Grid>
       )}
-    </section>
+    </Box>
   );
 }
 
-export function ReferralCreate({ showModal, setShowModal }) {
-  const { state, setState } = useContext(ObjectContext);
-  const { register, handleSubmit, setValue } = useForm(); //, watch, errors, reset
-  const [error, setError] = useState(false);
-  const [success, setSuccess] = useState(false);
-  const [success1, setSuccess1] = useState(false);
-  const [success2, setSuccess2] = useState(false);
-  const [message, setMessage] = useState('');
-  const [clientId, setClientId] = useState();
-  const [locationId, setLocationId] = useState();
-  const [practionerId, setPractionerId] = useState();
-  const [type, setType] = useState();
-  // eslint-disable-next-line
-  const [facility, setFacility] = useState();
-  const ClientServ = client.service('appointments');
-  //const navigate=useNavigate()
-  const { user } = useContext(UserContext); //,setUser
-  // eslint-disable-next-line
-  const [currentUser, setCurrentUser] = useState();
-  const [selectedClient, setSelectedClient] = useState();
-  const [selectedAppointment, setSelectedAppointment] = useState();
-  // const [appointment_reason,setAppointment_reason]= useState()
-  const [appointment_status, setAppointment_status] = useState('');
-  const [appointment_type, setAppointment_type] = useState('');
-  const [billingModal, setBillingModal] = useState(false);
+// export function ReferralCreate({ showModal, setShowModal }) {
+//   const { state, setState } = useContext(ObjectContext);
+//   const { register, handleSubmit, setValue } = useForm(); //, watch, errors, reset
+//   const [error, setError] = useState(false);
+//   const [success, setSuccess] = useState(false);
+//   const [success1, setSuccess1] = useState(false);
+//   const [success2, setSuccess2] = useState(false);
+//   const [message, setMessage] = useState("");
+//   const [clientId, setClientId] = useState();
+//   const [locationId, setLocationId] = useState();
+//   const [practionerId, setPractionerId] = useState();
+//   const [type, setType] = useState();
+//   // eslint-disable-next-line
+//   const [facility, setFacility] = useState();
+//   const ClientServ = client.service("appointments");
+//   //const navigate=useNavigate()
+//   const { user } = useContext(UserContext); //,setUser
+//   // eslint-disable-next-line
+//   const [currentUser, setCurrentUser] = useState();
+//   const [selectedClient, setSelectedClient] = useState();
+//   const [selectedAppointment, setSelectedAppointment] = useState();
+//   // const [appointment_reason,setAppointment_reason]= useState()
+//   const [appointment_status, setAppointment_status] = useState("");
+//   const [appointment_type, setAppointment_type] = useState("");
+//   const [billingModal, setBillingModal] = useState(false);
 
-  const [chosen, setChosen] = useState();
-  const [chosen1, setChosen1] = useState();
-  const [chosen2, setChosen2] = useState();
-  const [openComplaint, setOpenComplaint] = useState(false);
-  const [openFindings, setOpenFindings] = useState(false);
-  const [patient, setPatient] = useState('');
+//   const [chosen, setChosen] = useState();
+//   const [chosen1, setChosen1] = useState();
+//   const [chosen2, setChosen2] = useState();
+//   const [openComplaint, setOpenComplaint] = useState(false);
+//   const [openFindings, setOpenFindings] = useState(false);
+//   const [patient, setPatient] = useState("");
 
-  const appClass = ['On-site', 'Teleconsultation', 'Home Visit'];
+//   const appClass = ["On-site", "Teleconsultation", "Home Visit"];
 
-  let appointee; //  =state.ClientModule.selectedClient
-  /*  const getSearchfacility=(obj)=>{
-        setValue("facility", obj._id,  {
-            shouldValidate: true,
-            shouldDirty: true
-        })
-    } */
-  const handleChangeType = async (e) => {
-    await setAppointment_type(e.target.value);
-  };
+//   let appointee; //  =state.ClientModule.selectedClient
+//   /*  const getSearchfacility=(obj)=>{
+//         setValue("facility", obj._id,  {
+//             shouldValidate: true,
+//             shouldDirty: true
+//         })
+//     } */
+//   const handleChangeType = async (e) => {
+//     await setAppointment_type(e.target.value);
+//   };
 
-  const handleChangeStatus = async (e) => {
-    await setAppointment_status(e.target.value);
-  };
+//   const handleChangeStatus = async (e) => {
+//     await setAppointment_status(e.target.value);
+//   };
 
-  const getSearchfacility = (obj) => {
-    setClientId(obj._id);
-    setChosen(obj);
-    //handleRow(obj)
-    if (!obj) {
-      //"clear stuff"
-      setClientId();
-      setChosen();
-    }
+//   const getSearchfacility = (obj) => {
+//     setClientId(obj._id);
+//     setChosen(obj);
+//     //handleRow(obj)
+//     if (!obj) {
+//       //"clear stuff"
+//       setClientId();
+//       setChosen();
+//     }
 
-    /*  setValue("facility", obj._id,  {
-            shouldValidate: true,
-            shouldDirty: true
-        }) */
-  };
-  const getSearchfacility1 = (obj) => {
-    setLocationId(obj._id);
-    setChosen1(obj);
+//     /*  setValue("facility", obj._id,  {
+//             shouldValidate: true,
+//             shouldDirty: true
+//         }) */
+//   };
+//   const getSearchfacility1 = (obj) => {
+//     setLocationId(obj._id);
+//     setChosen1(obj);
 
-    if (!obj) {
-      //"clear stuff"
-      setLocationId();
-      setChosen1();
-    }
-  };
-  const getSearchfacility2 = (obj) => {
-    setPractionerId(obj._id);
-    setChosen2(obj);
+//     if (!obj) {
+//       //"clear stuff"
+//       setLocationId();
+//       setChosen1();
+//     }
+//   };
+//   const getSearchfacility2 = (obj) => {
+//     setPractionerId(obj._id);
+//     setChosen2(obj);
 
-    if (!obj) {
-      //"clear stuff"
-      setPractionerId();
-      setChosen2();
-    }
-  };
+//     if (!obj) {
+//       //"clear stuff"
+//       setPractionerId();
+//       setChosen2();
+//     }
+//   };
 
-  useEffect(() => {
-    setCurrentUser(user);
-    //console.log(currentUser)
-    return () => {};
-  }, [user]);
+//   useEffect(() => {
+//     setCurrentUser(user);
+//     //console.log(currentUser)
+//     return () => {};
+//   }, [user]);
 
-  //check user for facility or get list of facility
-  useEffect(() => {
-    //setFacility(user.activeClient.FacilityId)//
-    if (!user.stacker) {
-      /*    console.log(currentUser)
-        setValue("facility", user.currentEmployee.facilityDetail._id,  {
-            shouldValidate: true,
-            shouldDirty: true
-        })  */
-    }
-  });
+//   //check user for facility or get list of facility
+//   useEffect(() => {
+//     //setFacility(user.activeClient.FacilityId)//
+//     if (!user.stacker) {
+//       /*    console.log(currentUser)
+//         setValue("facility", user.currentEmployee.facilityDetail._id,  {
+//             shouldValidate: true,
+//             shouldDirty: true
+//         })  */
+//     }
+//   });
 
-  const onSubmit = (data, e) => {
-    e.preventDefault();
-    setMessage('');
-    setError(false);
-    setSuccess(false);
-    setShowModal(false),
-      setState((prevstate) => ({
-        ...prevstate,
-        AppointmentModule: {
-          selectedAppointment: {},
-          show: 'list',
-        },
-      }));
+//   const onSubmit = (data, e) => {
+//     e.preventDefault();
+//     setMessage("");
+//     setError(false);
+//     setSuccess(false);
+//     setShowModal(false),
+//       setState((prevstate) => ({
+//         ...prevstate,
+//         AppointmentModule: {
+//           selectedAppointment: {},
+//           show: "list",
+//         },
+//       }));
 
-    // data.createdby=user._id
-    console.log(data);
-    if (user.currentEmployee) {
-      data.facility = user.currentEmployee.facilityDetail._id; // or from facility dropdown
-    }
-    data.locationId = locationId; //state.ClinicModule.selectedClinic._id
-    data.practitionerId = practionerId;
-    data.appointment_type = appointment_type;
-    // data.appointment_reason=appointment_reason
-    data.appointment_status = appointment_status;
-    data.clientId = clientId;
-    data.firstname = chosen.firstname;
-    data.middlename = chosen.middlename;
-    data.lastname = chosen.lastname;
-    data.dob = chosen.dob;
-    data.gender = chosen.gender;
-    data.phone = chosen.phone;
-    data.email = chosen.email;
-    data.practitioner_name = chosen2.firstname + ' ' + chosen2.lastname;
-    data.practitioner_profession = chosen2.profession;
-    data.practitioner_department = chosen2.department;
-    data.location_name = chosen1.name;
-    data.location_type = chosen1.locationType;
-    data.actions = [
-      {
-        action: appointment_status,
-        actor: user.currentEmployee._id,
-      },
-    ];
-    console.log(data);
+//     // data.createdby=user._id
+//     console.log(data);
+//     if (user.currentEmployee) {
+//       data.facility = user.currentEmployee.facilityDetail._id; // or from facility dropdown
+//     }
+//     data.locationId = locationId; //state.ClinicModule.selectedClinic._id
+//     data.practitionerId = practionerId;
+//     data.appointment_type = appointment_type;
+//     // data.appointment_reason=appointment_reason
+//     data.appointment_status = appointment_status;
+//     data.clientId = clientId;
+//     data.firstname = chosen.firstname;
+//     data.middlename = chosen.middlename;
+//     data.lastname = chosen.lastname;
+//     data.dob = chosen.dob;
+//     data.gender = chosen.gender;
+//     data.phone = chosen.phone;
+//     data.email = chosen.email;
+//     data.practitioner_name = chosen2.firstname + " " + chosen2.lastname;
+//     data.practitioner_profession = chosen2.profession;
+//     data.practitioner_department = chosen2.department;
+//     data.location_name = chosen1.name;
+//     data.location_type = chosen1.locationType;
+//     data.actions = [
+//       {
+//         action: appointment_status,
+//         actor: user.currentEmployee._id,
+//       },
+//     ];
+//     console.log(data);
 
-    ClientServ.create(data)
-      .then((res) => {
-        //console.log(JSON.stringify(res))
-        e.target.reset();
-        setAppointment_type('');
-        setAppointment_status('');
-        setClientId('');
-        setLocationId('');
-        /*  setMessage("Created Client successfully") */
-        setSuccess(true);
-        setSuccess1(true);
-        setSuccess2(true);
-        toast({
-          message:
-            'Appointment created succesfully, Kindly bill patient if required',
-          type: 'is-success',
-          dismissible: true,
-          pauseOnHover: true,
-        });
-        setSuccess(false);
-        setSuccess1(false);
-        setSuccess2(false);
-        // showBilling()
-      })
-      .catch((err) => {
-        toast({
-          message: 'Error creating Appointment ' + err,
-          type: 'is-danger',
-          dismissible: true,
-          pauseOnHover: true,
-        });
-      });
-  };
+//     ClientServ.create(data)
+//       .then((res) => {
+//         //console.log(JSON.stringify(res))
+//         e.target.reset();
+//         setAppointment_type("");
+//         setAppointment_status("");
+//         setClientId("");
+//         setLocationId("");
+//         /*  setMessage("Created Client successfully") */
+//         setSuccess(true);
+//         setSuccess1(true);
+//         setSuccess2(true);
+//         toast({
+//           message:
+//             "Appointment created succesfully, Kindly bill patient if required",
+//           type: "is-success",
+//           dismissible: true,
+//           pauseOnHover: true,
+//         });
+//         setSuccess(false);
+//         setSuccess1(false);
+//         setSuccess2(false);
+//         // showBilling()
+//       })
+//       .catch((err) => {
+//         toast({
+//           message: "Error creating Appointment " + err,
+//           type: "is-danger",
+//           dismissible: true,
+//           pauseOnHover: true,
+//         });
+//       });
+//   };
 
-  useEffect(() => {
-    getSearchfacility(state.ClientModule.selectedClient);
+//   useEffect(() => {
+//     getSearchfacility(state.ClientModule.selectedClient);
 
-    /* appointee=state.ClientModule.selectedClient 
-        console.log(appointee.firstname) */
-    return () => {};
-  }, [state.ClientModule.selectedClient]);
+//     /* appointee=state.ClientModule.selectedClient 
+//         console.log(appointee.firstname) */
+//     return () => {};
+//   }, [state.ClientModule.selectedClient]);
 
-  /*   const showBilling = () =>{
-        setBillingModal(true)
-       //history.push('/app/finance/billservice')
-        }
-        const  handlecloseModal1 = () =>{
-            setBillingModal(false)
-            }
+//   /*   const showBilling = () =>{
+//         setBillingModal(true)
+//        //history.push('/app/finance/billservice')
+//         }
+//         const  handlecloseModal1 = () =>{
+//             setBillingModal(false)
+//             }
 
 
-            const handleRow= async(Client)=>{
-              //  await setSelectedClient(Client)
-                const    newClientModule={
-                    selectedClient:Client,
-                    show :'detail'
-                }
-               await setState((prevstate)=>({...prevstate, ClientModule:newClientModule}))
-            } */
-  const CustomSelectData = [
-    {
-      label: 'Today',
-      value: 'today',
-    },
-  ];
+//             const handleRow= async(Client)=>{
+//               //  await setSelectedClient(Client)
+//                 const    newClientModule={
+//                     selectedClient:Client,
+//                     show :'detail'
+//                 }
+//                await setState((prevstate)=>({...prevstate, ClientModule:newClientModule}))
+//             } */
+//   const CustomSelectData = [
+//     {
+//       label: "Today",
+//       value: "today",
+//     },
+//   ];
 
-  const dummyData = [
-    {
-      complaint: 'Fever',
-      duration: '2 days',
-    },
-  ];
-  const complaintSchema = [
-    {
-      name: 'S/N',
-      key: 'sn',
-      description: 'SN',
-      selector: (row) => row.sn,
-      sortable: true,
-      inputType: 'HIDDEN',
-    },
-    {
-      name: 'Complaint',
-      key: 'complaint',
-      description: 'Complaint',
-      selector: (row) => row.complaint,
-      sortable: true,
-      inputType: 'TEXT',
-    },
-    {
-      name: 'Duration',
-      key: 'duration',
-      description: 'Duration',
-      selector: (row) => row.duration,
-      sortable: true,
-      inputType: 'TEXT',
-    },
-  ];
-  const dummyData2 = [
-    {
-      provisional: 'Fever',
-      procedure: 'Test',
-      service: 'Test',
-    },
-  ];
-  const findingsSchema = [
-    {
-      name: 'S/N',
-      key: 'sn',
-      description: 'SN',
-      selector: (row) => row.sn,
-      sortable: true,
-      inputType: 'HIDDEN',
-    },
-    {
-      name: 'Provisional Diagnosis',
-      key: 'provisional',
-      description: 'Provisional Diagnosis',
-      selector: (row) => row.provisional,
-      sortable: true,
-      inputType: 'TEXT',
-    },
-    {
-      name: 'Procedure',
-      key: 'procedure',
-      description: 'Planned Procedure',
-      selector: (row) => row.procedure,
-      sortable: true,
-      inputType: 'TEXT',
-    },
-    {
-      name: 'Service',
-      key: 'service',
-      description: 'Planned Service',
-      selector: (row) => row.service,
-      sortable: true,
-      inputType: 'TEXT',
-    },
-  ];
+//   const dummyData = [
+//     {
+//       complaint: "Fever",
+//       duration: "2 days",
+//     },
+//   ];
+//   const complaintSchema = [
+//     {
+//       name: "S/N",
+//       key: "sn",
+//       description: "SN",
+//       selector: (row) => row.sn,
+//       sortable: true,
+//       inputType: "HIDDEN",
+//     },
+//     {
+//       name: "Complaint",
+//       key: "complaint",
+//       description: "Complaint",
+//       selector: (row) => row.complaint,
+//       sortable: true,
+//       inputType: "TEXT",
+//     },
+//     {
+//       name: "Duration",
+//       key: "duration",
+//       description: "Duration",
+//       selector: (row) => row.duration,
+//       sortable: true,
+//       inputType: "TEXT",
+//     },
+//   ];
+//   const dummyData2 = [
+//     {
+//       provisional: "Fever",
+//       procedure: "Test",
+//       service: "Test",
+//     },
+//   ];
+//   const findingsSchema = [
+//     {
+//       name: "S/N",
+//       key: "sn",
+//       description: "SN",
+//       selector: (row) => row.sn,
+//       sortable: true,
+//       inputType: "HIDDEN",
+//     },
+//     {
+//       name: "Provisional Diagnosis",
+//       key: "provisional",
+//       description: "Provisional Diagnosis",
+//       selector: (row) => row.provisional,
+//       sortable: true,
+//       inputType: "TEXT",
+//     },
+//     {
+//       name: "Procedure",
+//       key: "procedure",
+//       description: "Planned Procedure",
+//       selector: (row) => row.procedure,
+//       sortable: true,
+//       inputType: "TEXT",
+//     },
+//     {
+//       name: "Service",
+//       key: "service",
+//       description: "Planned Service",
+//       selector: (row) => row.service,
+//       sortable: true,
+//       inputType: "TEXT",
+//     },
+//   ];
 
-  return (
-    <>
-      <div
-        className="card "
-        style={{
-          margin: '0 auto',
-          width: '98%',
-          height: 'calc(100vh - 90px)',
-          overflow: 'scroll',
-        }}
-      >
-        <form onSubmit={handleSubmit(onSubmit)}>
-          <Grid container spacing={2}>
-            <Grid item xs={12} sm={6}>
-              <ModalHeader text={'Referrals'} />
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <Box sx={{ display: 'flex', justifyContent: 'flex-end' }}>
-                <GlobalCustomButton
-                  onClick={() => {
-                    setShowModal(0),
-                      setState((prevstate) => ({
-                        ...prevstate,
-                        AppointmentModule: {
-                          selectedAppointment: {},
-                          show: 'list',
-                        },
-                      }));
-                  }}
-                  text={'Back'}
-                  color="warning"
-                  customStyles={{ marginRight: '.8rem' }}
-                />
-                <GlobalCustomButton
-                  type="submit"
-                  text={'Save'}
-                  color="success"
-                />
-              </Box>
-            </Grid>
-          </Grid>
-          <Grid container spacing={2}>
-            <Grid item xs={12} sm={12}>
-              <RadioButton
-                name="patient"
-                title="Patient"
-                options={[
-                  {
-                    label: 'Out Patient',
-                    value: 'Out Patient',
-                  },
-                  {
-                    label: 'In Patient',
-                    value: 'In Patient',
-                  },
-                ]}
-                onChange={(e) => setPatient(e.target.value)}
-              />
-            </Grid>
-          </Grid>
-          <Grid container spacing={2} mt={1}>
-            <Grid item xs={12} sm={4}>
-              <Input name="patientName" label="Search Beneficiary" />
-            </Grid>
-            <Grid item xs={12} sm={4}>
-              <FacilitySearch
-                getSearchfacility={getSearchfacility}
-                clear={success}
-                label="Destination Facility"
-              />
-            </Grid>
-            <Grid item xs={12} sm={4}>
-              <Input
-                name="patientName"
-                label="Referring Facility"
-                value={'Test Organization'}
-              />
-            </Grid>
-            {patient === 'In Patient' && (
-              <Grid item xs={12} sm={4}>
-                <BasicDatePicker
-                  name="addmissionDate"
-                  label="Date of Admission"
-                />
-              </Grid>
-            )}
-            {patient === 'In Patient' && (
-              <Grid item xs={12} sm={4}>
-                <BasicDatePicker
-                  name="dischargeDate"
-                  label="Date of Discharge"
-                />
-              </Grid>
-            )}
-          </Grid>
+//   return (
+//     <>
+//       <div
+//         className="card "
+//         style={{
+//           margin: "0 auto",
+//           width: "98%",
+//           height: "calc(100vh - 90px)",
+//           overflow: "scroll",
+//         }}
+//       >
+//         <form onSubmit={handleSubmit(onSubmit)}>
+//           <Grid container spacing={2}>
+//             <Grid item xs={12} sm={6}>
+//               <ModalHeader text={"Referrals"} />
+//             </Grid>
+//             <Grid item xs={12} sm={6}>
+//               <Box sx={{ display: "flex", justifyContent: "flex-end" }}>
+//                 <GlobalCustomButton
+//                   onClick={() => {
+//                     setShowModal(0),
+//                       setState((prevstate) => ({
+//                         ...prevstate,
+//                         AppointmentModule: {
+//                           selectedAppointment: {},
+//                           show: "list",
+//                         },
+//                       }));
+//                   }}
+//                   text={"Back"}
+//                   color="warning"
+//                   customStyles={{ marginRight: ".8rem" }}
+//                 />
+//                 <GlobalCustomButton
+//                   type="submit"
+//                   text={"Save"}
+//                   color="success"
+//                 />
+//               </Box>
+//             </Grid>
+//           </Grid>
+//           <Grid container spacing={2}>
+//             <Grid item xs={12} sm={12}>
+//               <RadioButton
+//                 name="patient"
+//                 title="Patient"
+//                 options={[
+//                   {
+//                     label: "Out Patient",
+//                     value: "Out Patient",
+//                   },
+//                   {
+//                     label: "In Patient",
+//                     value: "In Patient",
+//                   },
+//                 ]}
+//                 onChange={(e) => setPatient(e.target.value)}
+//               />
+//             </Grid>
+//           </Grid>
+//           <Grid container spacing={2} mt={1}>
+//             <Grid item xs={12} sm={4}>
+//               <Input name="patientName" label="Search Beneficiary" />
+//             </Grid>
+//             <Grid item xs={12} sm={4}>
+//               <FacilitySearch
+//                 getSearchfacility={getSearchfacility}
+//                 clear={success}
+//                 label="Destination Facility"
+//               />
+//             </Grid>
+//             <Grid item xs={12} sm={4}>
+//               <Input
+//                 name="patientName"
+//                 label="Referring Facility"
+//                 value={"Test Organization"}
+//               />
+//             </Grid>
+//             {patient === "In Patient" && (
+//               <Grid item xs={12} sm={4}>
+//                 <BasicDatePicker
+//                   name="addmissionDate"
+//                   label="Date of Admission"
+//                 />
+//               </Grid>
+//             )}
+//             {patient === "In Patient" && (
+//               <Grid item xs={12} sm={4}>
+//                 <BasicDatePicker
+//                   name="dischargeDate"
+//                   label="Date of Discharge"
+//                 />
+//               </Grid>
+//             )}
+//           </Grid>
 
-          <Grid container spacing={2} my={2}>
-            <Grid item xs={12} sm={6}>
-              <FormsHeaderText text={'Clinical Information'} />
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <GlobalCustomButton
-                customStyles={{
-                  float: 'right',
-                }}
-                text={'Add Complaint'}
-                color="primary"
-                variant="outlined"
-                onClick={() => setOpenComplaint(true)}
-              />
-            </Grid>
-          </Grid>
-          <CustomTable
-            title={''}
-            columns={complaintSchema}
-            data={dummyData}
-            pointerOnHover
-            highlightOnHover
-            striped
-          />
+//           <Grid container spacing={2} my={2}>
+//             <Grid item xs={12} sm={6}>
+//               <FormsHeaderText text={"Clinical Information"} />
+//             </Grid>
+//             <Grid item xs={12} sm={6}>
+//               <GlobalCustomButton
+//                 customStyles={{
+//                   float: "right",
+//                 }}
+//                 text={"Add Complaint"}
+//                 color="primary"
+//                 variant="outlined"
+//                 onClick={() => setOpenComplaint(true)}
+//               />
+//             </Grid>
+//           </Grid>
+//           <CustomTable
+//             title={""}
+//             columns={complaintSchema}
+//             data={dummyData}
+//             pointerOnHover
+//             highlightOnHover
+//             striped
+//           />
 
-          <Grid container spacing={2} my={2}>
-            <Grid item xs={12} sm={12}>
-              <FormsHeaderText
-                txt={'Clinic Findings'}
-                color={'#0064CC'}
-                type={'p'}
-                bold={'700'}
-                size={'18px'}
-              />
-            </Grid>
-          </Grid>
+//           <Grid container spacing={2} my={2}>
+//             <Grid item xs={12} sm={12}>
+//               <FormsHeaderText
+//                 txt={"Clinic Findings"}
+//                 color={"#0064CC"}
+//                 type={"p"}
+//                 bold={"700"}
+//                 size={"18px"}
+//               />
+//             </Grid>
+//           </Grid>
 
-          <Grid container spacing={2} my={2}>
-            <Grid item xs={12} sm={6}>
-              <FormsHeaderText text={'Clinical Findings'} />
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <GlobalCustomButton
-                customStyles={{
-                  float: 'right',
-                }}
-                text={'Add Findings'}
-                color="primary"
-                variant="outlined"
-                onClick={() => setOpenFindings(true)}
-              />
-            </Grid>
-          </Grid>
-          <CustomTable
-            title={''}
-            columns={findingsSchema}
-            data={dummyData2}
-            pointerOnHover
-            highlightOnHover
-            striped
-          />
+//           <Grid container spacing={2} my={2}>
+//             <Grid item xs={12} sm={6}>
+//               <FormsHeaderText text={"Clinical Findings"} />
+//             </Grid>
+//             <Grid item xs={12} sm={6}>
+//               <GlobalCustomButton
+//                 customStyles={{
+//                   float: "right",
+//                 }}
+//                 text={"Add Findings"}
+//                 color="primary"
+//                 variant="outlined"
+//                 onClick={() => setOpenFindings(true)}
+//               />
+//             </Grid>
+//           </Grid>
+//           <CustomTable
+//             title={""}
+//             columns={findingsSchema}
+//             data={dummyData2}
+//             pointerOnHover
+//             highlightOnHover
+//             striped
+//           />
 
-          <Grid container spacing={2}>
-            <Grid item xs={12} sm={12}>
-              <Textarea
-                placeholder="Type your message here"
-                name="reason"
-                type="text"
-                label="Reason for Request"
-              />
-            </Grid>
-          </Grid>
+//           <Grid container spacing={2}>
+//             <Grid item xs={12} sm={12}>
+//               <Textarea
+//                 placeholder="Type your message here"
+//                 name="reason"
+//                 type="text"
+//                 label="Reason for Request"
+//               />
+//             </Grid>
+//           </Grid>
 
-          <Grid container spacing={2} my={1}>
-            <Grid item xs={12} sm={6}>
-              <Input
-                name="physicianName"
-                label="Physician's Name"
-                type="text"
-              />
-            </Grid>
-          </Grid>
-        </form>
-      </div>
-      {openComplaint && (
-        <ModalBox
-          open={openComplaint}
-          onClose={() => setOpenComplaint(false)}
-          header="Add Complaint"
-        >
-          <Grid container spacing={2}>
-            <Grid item xs={12} sm={12}>
-              <Input name="complaints" label="Complaints" />
-            </Grid>
-            <Grid item xs={12} sm={12}>
-              <Input name="duration" label="Duration" />
-            </Grid>
-            <Grid item xs={12} sm={12}>
-              <GlobalCustomButton text={'Add'} color="success" />
-            </Grid>
-          </Grid>
-        </ModalBox>
-      )}
-      {openFindings && (
-        <ModalBox
-          open={openFindings}
-          onClose={() => setOpenFindings(false)}
-          header="Add Findings"
-        >
-          <Grid container spacing={2}>
-            <Grid item xs={12} sm={12}>
-              <AutoCompleteBox
-                label="Provisional Diagnosis"
-                name="provisionalDiagnosis"
-                options={[
-                  { value: 'Fever', label: 'Fever' },
-                  { value: 'Cough', label: 'Cough' },
-                  { value: 'Headache', label: 'Headache' },
-                  { value: 'Body Pain', label: 'Body Pain' },
-                  { value: 'Diarrhea', label: 'Diarrhea' },
-                  { value: 'Vomiting', label: 'Vomiting' },
-                ]}
-              />
-            </Grid>
-            <Grid item xs={12} sm={12}>
-              <Input label="Planned Procedure" />
-            </Grid>
-            <Grid item xs={12} sm={12}>
-              <Input label="Planned Service" />
-            </Grid>
-            <Grid item xs={12} sm={12}>
-              <GlobalCustomButton text={'Add'} color="success" />
-            </Grid>
-          </Grid>
-        </ModalBox>
-      )}
-    </>
-  );
-}
+//           <Grid container spacing={2} my={1}>
+//             <Grid item xs={12} sm={6}>
+//               <Input
+//                 name="physicianName"
+//                 label="Physician's Name"
+//                 type="text"
+//               />
+//             </Grid>
+//           </Grid>
+//         </form>
+//       </div>
+//       {openComplaint && (
+//         <ModalBox
+//           open={openComplaint}
+//           onClose={() => setOpenComplaint(false)}
+//           header="Add Complaint"
+//         >
+//           <Grid container spacing={2}>
+//             <Grid item xs={12} sm={12}>
+//               <Input name="complaints" label="Complaints" />
+//             </Grid>
+//             <Grid item xs={12} sm={12}>
+//               <Input name="duration" label="Duration" />
+//             </Grid>
+//             <Grid item xs={12} sm={12}>
+//               <GlobalCustomButton text={"Add"} color="success" />
+//             </Grid>
+//           </Grid>
+//         </ModalBox>
+//       )}
+//       {openFindings && (
+//         <ModalBox
+//           open={openFindings}
+//           onClose={() => setOpenFindings(false)}
+//           header="Add Findings"
+//         >
+//           <Grid container spacing={2}>
+//             <Grid item xs={12} sm={12}>
+//               <AutoCompleteBox
+//                 label="Provisional Diagnosis"
+//                 name="provisionalDiagnosis"
+//                 options={[
+//                   { value: "Fever", label: "Fever" },
+//                   { value: "Cough", label: "Cough" },
+//                   { value: "Headache", label: "Headache" },
+//                   { value: "Body Pain", label: "Body Pain" },
+//                   { value: "Diarrhea", label: "Diarrhea" },
+//                   { value: "Vomiting", label: "Vomiting" },
+//                 ]}
+//               />
+//             </Grid>
+//             <Grid item xs={12} sm={12}>
+//               <Input label="Planned Procedure" />
+//             </Grid>
+//             <Grid item xs={12} sm={12}>
+//               <Input label="Planned Service" />
+//             </Grid>
+//             <Grid item xs={12} sm={12}>
+//               <GlobalCustomButton text={"Add"} color="success" />
+//             </Grid>
+//           </Grid>
+//         </ModalBox>
+//       )}
+//     </>
+//   );
+// }
 
-export function ReferralList({ showModal, setShowModal }) {
-  // const { register, handleSubmit, watch, errors } = useForm();
-  // eslint-disable-next-line
-  const [error, setError] = useState(false);
-  // eslint-disable-next-line
-  const [success, setSuccess] = useState(false);
-  // eslint-disable-next-line
-  const [message, setMessage] = useState('');
-  const ClientServ = client.service('appointments');
-  //const navigate=useNavigate()
-  // const {user,setUser} = useContext(UserContext)
-  const [facilities, setFacilities] = useState([]);
-  // eslint-disable-next-line
-  const [selectedClient, setSelectedClient] = useState(); //
-  // eslint-disable-next-line
-  const { state, setState } = useContext(ObjectContext);
-  // eslint-disable-next-line
-  const { user, setUser } = useContext(UserContext);
-  const [startDate, setStartDate] = useState(new Date());
-  const [selectedAppointment, setSelectedAppointment] = useState();
-  const [loading, setLoading] = useState(false);
-  const [value, setValue] = useState('list');
+// export function ReferralList({ showModal, setShowModal }) {
+//   // const { register, handleSubmit, watch, errors } = useForm();
+//   // eslint-disable-next-line
+//   const [error, setError] = useState(false);
+//   // eslint-disable-next-line
+//   const [success, setSuccess] = useState(false);
+//   // eslint-disable-next-line
+//   const [message, setMessage] = useState('');
+//   const ClientServ = client.service('appointments');
+//   //const navigate=useNavigate()
+//   // const {user,setUser} = useContext(UserContext)
+//   const [facilities, setFacilities] = useState([]);
+//   // eslint-disable-next-line
+//   const [selectedClient, setSelectedClient] = useState(); //
+//   // eslint-disable-next-line
+//   const { state, setState } = useContext(ObjectContext);
+//   // eslint-disable-next-line
+//   const { user, setUser } = useContext(UserContext);
+//   const [startDate, setStartDate] = useState(new Date());
+//   const [selectedAppointment, setSelectedAppointment] = useState();
+//   const [loading, setLoading] = useState(false);
+//   const [value, setValue] = useState('list');
 
-  const handleCreateNew = async () => {
-    const newClientModule = {
-      selectedAppointment: {},
-      show: 'create',
-    };
-    await setState((prevstate) => ({
-      ...prevstate,
-      AppointmentModule: newClientModule,
-    }));
-    //console.log(state)
-    const newClient = {
-      selectedClient: {},
-      show: 'create',
-    };
-    await setState((prevstate) => ({ ...prevstate, ClientModule: newClient }));
-    setShowModal(1);
-  };
+//   const handleCreateNew = async () => {
+//     const newClientModule = {
+//       selectedAppointment: {},
+//       show: 'create',
+//     };
+//     await setState((prevstate) => ({
+//       ...prevstate,
+//       AppointmentModule: newClientModule,
+//     }));
+//     //console.log(state)
+//     const newClient = {
+//       selectedClient: {},
+//       show: 'create',
+//     };
+//     await setState((prevstate) => ({ ...prevstate, ClientModule: newClient }));
+//     setShowModal(1);
+//   };
 
-  const handleRow = async (Client) => {
-    setShowModal(2);
-    await setSelectedAppointment(Client);
-    const newClientModule = {
-      selectedAppointment: Client,
-      show: 'detail',
-    };
-    await setState((prevstate) => ({
-      ...prevstate,
-      AppointmentModule: newClientModule,
-    }));
-  };
-  //console.log(state.employeeLocation)
+//   const handleRow = async (Client) => {
+//     setShowModal(2);
+//     await setSelectedAppointment(Client);
+//     const newClientModule = {
+//       selectedAppointment: Client,
+//       show: 'detail',
+//     };
+//     await setState((prevstate) => ({
+//       ...prevstate,
+//       AppointmentModule: newClientModule,
+//     }));
+//   };
+//   //console.log(state.employeeLocation)
 
-  const handleSearch = (val) => {
-    const field = 'firstname';
-    //  console.log(val)
+//   const handleSearch = (val) => {
+//     const field = 'firstname';
+//     //  console.log(val)
 
-    let query = {
-      $or: [
-        {
-          firstname: {
-            $regex: val,
-            $options: 'i',
-          },
-        },
-        {
-          lastname: {
-            $regex: val,
-            $options: 'i',
-          },
-        },
-        {
-          middlename: {
-            $regex: val,
-            $options: 'i',
-          },
-        },
-        {
-          phone: {
-            $regex: val,
-            $options: 'i',
-          },
-        },
-        {
-          appointment_type: {
-            $regex: val,
-            $options: 'i',
-          },
-        },
-        {
-          appointment_status: {
-            $regex: val,
-            $options: 'i',
-          },
-        },
-        {
-          appointment_reason: {
-            $regex: val,
-            $options: 'i',
-          },
-        },
-        {
-          location_type: {
-            $regex: val,
-            $options: 'i',
-          },
-        },
-        {
-          location_name: {
-            $regex: val,
-            $options: 'i',
-          },
-        },
-        {
-          practitioner_department: {
-            $regex: val,
-            $options: 'i',
-          },
-        },
-        {
-          practitioner_profession: {
-            $regex: val,
-            $options: 'i',
-          },
-        },
-        {
-          practitioner_name: {
-            $regex: val,
-            $options: 'i',
-          },
-        },
-      ],
-      facility: user.currentEmployee.facilityDetail._id, // || "",
-      $limit: 20,
-      $sort: {
-        createdAt: -1,
-      },
-    };
-    if (state.employeeLocation.locationType !== 'Front Desk') {
-      query.locationId = state.employeeLocation.locationId;
-    }
+//     let query = {
+//       $or: [
+//         {
+//           firstname: {
+//             $regex: val,
+//             $options: 'i',
+//           },
+//         },
+//         {
+//           lastname: {
+//             $regex: val,
+//             $options: 'i',
+//           },
+//         },
+//         {
+//           middlename: {
+//             $regex: val,
+//             $options: 'i',
+//           },
+//         },
+//         {
+//           phone: {
+//             $regex: val,
+//             $options: 'i',
+//           },
+//         },
+//         {
+//           appointment_type: {
+//             $regex: val,
+//             $options: 'i',
+//           },
+//         },
+//         {
+//           appointment_status: {
+//             $regex: val,
+//             $options: 'i',
+//           },
+//         },
+//         {
+//           appointment_reason: {
+//             $regex: val,
+//             $options: 'i',
+//           },
+//         },
+//         {
+//           location_type: {
+//             $regex: val,
+//             $options: 'i',
+//           },
+//         },
+//         {
+//           location_name: {
+//             $regex: val,
+//             $options: 'i',
+//           },
+//         },
+//         {
+//           practitioner_department: {
+//             $regex: val,
+//             $options: 'i',
+//           },
+//         },
+//         {
+//           practitioner_profession: {
+//             $regex: val,
+//             $options: 'i',
+//           },
+//         },
+//         {
+//           practitioner_name: {
+//             $regex: val,
+//             $options: 'i',
+//           },
+//         },
+//       ],
+//       facility: user.currentEmployee.facilityDetail._id, // || "",
+//       $limit: 20,
+//       $sort: {
+//         createdAt: -1,
+//       },
+//     };
+//     if (state.employeeLocation.locationType !== 'Front Desk') {
+//       query.locationId = state.employeeLocation.locationId;
+//     }
 
-    ClientServ.find({ query: query })
-      .then((res) => {
-        console.log(res);
-        setFacilities(res.data);
-        setMessage(' Client  fetched successfully');
-        setSuccess(true);
-      })
-      .catch((err) => {
-        console.log(err);
-        setMessage('Error fetching Client, probable network issues ' + err);
-        setError(true);
-      });
-  };
+//     ClientServ.find({ query: query })
+//       .then((res) => {
+//         console.log(res);
+//         setFacilities(res.data);
+//         setMessage(' Client  fetched successfully');
+//         setSuccess(true);
+//       })
+//       .catch((err) => {
+//         console.log(err);
+//         setMessage('Error fetching Client, probable network issues ' + err);
+//         setError(true);
+//       });
+//   };
 
-  const getFacilities = async () => {
-    console.log(user);
-    if (user.currentEmployee) {
-      let stuff = {
-        facility: user.currentEmployee.facilityDetail._id,
-        // locationId:state.employeeLocation.locationId,
-        $limit: 100,
-        $sort: {
-          createdAt: -1,
-        },
-      };
-      // if (state.employeeLocation.locationType !== "Front Desk") {
-      //   stuff.locationId = state.employeeLocation.locationId;
-      // }
+//   const getFacilities = async () => {
+//     console.log(user);
+//     if (user.currentEmployee) {
+//       let stuff = {
+//         facility: user.currentEmployee.facilityDetail._id,
+//         // locationId:state.employeeLocation.locationId,
+//         $limit: 100,
+//         $sort: {
+//           createdAt: -1,
+//         },
+//       };
+//       // if (state.employeeLocation.locationType !== "Front Desk") {
+//       //   stuff.locationId = state.employeeLocation.locationId;
+//       // }
 
-      const findClient = await ClientServ.find({ query: stuff });
+//       const findClient = await ClientServ.find({ query: stuff });
 
-      await setFacilities(findClient.data);
-      console.log(findClient.data);
-    } else {
-      if (user.stacker) {
-        const findClient = await ClientServ.find({
-          query: {
-            $limit: 100,
-            $sort: {
-              createdAt: -1,
-            },
-          },
-        });
+//       await setFacilities(findClient.data);
+//       console.log(findClient.data);
+//     } else {
+//       if (user.stacker) {
+//         const findClient = await ClientServ.find({
+//           query: {
+//             $limit: 100,
+//             $sort: {
+//               createdAt: -1,
+//             },
+//           },
+//         });
 
-        await setFacilities(findClient.data);
-      }
-    }
-  };
+//         await setFacilities(findClient.data);
+//       }
+//     }
+//   };
 
-  useEffect(() => {
-    if (user) {
-      handleCalendarClose();
-    } else {
-      /* const localUser= localStorage.getItem("user")
-                    const user1=JSON.parse(localUser)
-                    console.log(localUser)
-                    console.log(user1)
-                    fetchUser(user1)
-                    console.log(user)
-                    getFacilities(user) */
-    }
-    ClientServ.on('created', (obj) => handleCalendarClose());
-    ClientServ.on('updated', (obj) => handleCalendarClose());
-    ClientServ.on('patched', (obj) => handleCalendarClose());
-    ClientServ.on('removed', (obj) => handleCalendarClose());
-    const newClient = {
-      selectedClient: {},
-      show: 'create',
-    };
-    setState((prevstate) => ({ ...prevstate, ClientModule: newClient }));
-    return () => {};
-  }, []);
-  const handleCalendarClose = async () => {
-    let query = {
-      start_time: {
-        $gt: subDays(startDate, 1),
-        $lt: addDays(startDate, 1),
-      },
-      facility: user.currentEmployee.facilityDetail._id,
+//   useEffect(() => {
+//     if (user) {
+//       handleCalendarClose();
+//     } else {
+//       /* const localUser= localStorage.getItem("user")
+//                     const user1=JSON.parse(localUser)
+//                     console.log(localUser)
+//                     console.log(user1)
+//                     fetchUser(user1)
+//                     console.log(user)
+//                     getFacilities(user) */
+//     }
+//     ClientServ.on('created', (obj) => handleCalendarClose());
+//     ClientServ.on('updated', (obj) => handleCalendarClose());
+//     ClientServ.on('patched', (obj) => handleCalendarClose());
+//     ClientServ.on('removed', (obj) => handleCalendarClose());
+//     const newClient = {
+//       selectedClient: {},
+//       show: 'create',
+//     };
+//     setState((prevstate) => ({ ...prevstate, ClientModule: newClient }));
+//     return () => {};
+//   }, []);
+//   const handleCalendarClose = async () => {
+//     let query = {
+//       start_time: {
+//         $gt: subDays(startDate, 1),
+//         $lt: addDays(startDate, 1),
+//       },
+//       facility: user.currentEmployee.facilityDetail._id,
 
-      $limit: 100,
-      $sort: {
-        createdAt: -1,
-      },
-    };
-    // if (state.employeeLocation.locationType !== "Front Desk") {
-    //   query.locationId = state.employeeLocation.locationId;
-    // }
+//       $limit: 100,
+//       $sort: {
+//         createdAt: -1,
+//       },
+//     };
+//     // if (state.employeeLocation.locationType !== "Front Desk") {
+//     //   query.locationId = state.employeeLocation.locationId;
+//     // }
 
-    const findClient = await ClientServ.find({ query: query });
+//     const findClient = await ClientServ.find({ query: query });
 
-    await setFacilities(findClient.data);
-  };
+//     await setFacilities(findClient.data);
+//   };
 
-  const handleDate = async (date) => {
-    setStartDate(date);
-  };
+//   const handleDate = async (date) => {
+//     setStartDate(date);
+//   };
 
-  useEffect(() => {
-    if (!!startDate) {
-      handleCalendarClose();
-    } else {
-      getFacilities();
-    }
+//   useEffect(() => {
+//     if (!!startDate) {
+//       handleCalendarClose();
+//     } else {
+//       getFacilities();
+//     }
 
-    return () => {};
-  }, [startDate]);
-  //todo: pagination and vertical scroll bar
+//     return () => {};
+//   }, [startDate]);
+//   //todo: pagination and vertical scroll bar
 
-  const onRowClicked = () => {};
+//   const onRowClicked = () => {};
 
-  const mapFacilities = () => {
-    let mapped = [];
-    facilities.map((facility, i) => {
-      mapped.push({
-        title: facility?.firstname + ' ' + facility?.lastname,
-        start: format(new Date(facility?.start_time), 'yyyy-MM-ddTHH:mm'),
-        end: facility?.end_time,
-        id: i,
-      });
-    });
-    return mapped;
-  };
-  const activeStyle = {
-    backgroundColor: '#0064CC29',
-    border: 'none',
-    padding: '0 .8rem',
-  };
+//   const mapFacilities = () => {
+//     let mapped = [];
+//     facilities.map((facility, i) => {
+//       mapped.push({
+//         title: facility?.firstname + ' ' + facility?.lastname,
+//         start: format(new Date(facility?.start_time), 'yyyy-MM-ddTHH:mm'),
+//         end: facility?.end_time,
+//         id: i,
+//       });
+//     });
+//     return mapped;
+//   };
+//   const activeStyle = {
+//     backgroundColor: '#0064CC29',
+//     border: 'none',
+//     padding: '0 .8rem',
+//   };
 
-  const dummyData = [
-    {
-      date: '27/10/21',
-      patients_name: 'Tejiri Tabor',
-      policy_id: '234.75.43.01',
-      referral_code: '324234 - AC',
-      referral_provider: 'Creek Hospital',
-      destination_provider: 'Creek Hospital',
-      status: 'Approved',
-      reason_for_request: 'Lorem ipsum dolor',
-    },
-    {
-      date: '27/10/21',
-      patients_name: 'Tejiri Tabor',
-      policy_id: '234.75.43.01',
-      referral_code: '324234 - AC',
-      referral_provider: 'Creek Hospital',
-      destination_provider: 'Creek Hospital',
-      status: 'Approved',
-      reason_for_request: 'Lorem ipsum dolor',
-    },
-    {
-      date: '27/10/21',
-      patients_name: 'Tejiri Tabor',
-      policy_id: '234.75.43.01',
-      referral_code: '324234 - AC',
-      referral_provider: 'Creek Hospital',
-      destination_provider: 'Creek Hospital',
-      status: 'Approved',
-      reason_for_request: 'Lorem ipsum dolor',
-    },
+//   const dummyData = [
+//     {
+//       date: '27/10/21',
+//       patients_name: 'Tejiri Tabor',
+//       policy_id: '234.75.43.01',
+//       referral_code: '324234 - AC',
+//       referral_provider: 'Creek Hospital',
+//       destination_provider: 'Creek Hospital',
+//       status: 'Approved',
+//       reason_for_request: 'Lorem ipsum dolor',
+//     },
+//     {
+//       date: '27/10/21',
+//       patients_name: 'Tejiri Tabor',
+//       policy_id: '234.75.43.01',
+//       referral_code: '324234 - AC',
+//       referral_provider: 'Creek Hospital',
+//       destination_provider: 'Creek Hospital',
+//       status: 'Approved',
+//       reason_for_request: 'Lorem ipsum dolor',
+//     },
+//     {
+//       date: '27/10/21',
+//       patients_name: 'Tejiri Tabor',
+//       policy_id: '234.75.43.01',
+//       referral_code: '324234 - AC',
+//       referral_provider: 'Creek Hospital',
+//       destination_provider: 'Creek Hospital',
+//       status: 'Approved',
+//       reason_for_request: 'Lorem ipsum dolor',
+//     },
 
-    {
-      date: '27/10/21',
-      patients_name: 'Tejiri Tabor',
-      policy_id: '234.75.43.01',
-      referral_code: '324234 - AC',
-      referral_provider: 'Creek Hospital',
-      destination_provider: 'Creek Hospital',
-      status: 'Approved',
-      reason_for_request: 'Lorem ipsum dolor',
-    },
-    {
-      date: '27/10/21',
-      patients_name: 'Tejiri Tabor',
-      policy_id: '234.75.43.01',
-      referral_code: '324234 - AC',
-      referral_provider: 'Creek Hospital',
-      destination_provider: 'Creek Hospital',
-      status: 'Approved',
-      reason_for_request: 'Lorem ipsum dolor',
-    },
-    {
-      date: '27/10/21',
-      patients_name: 'Tejiri Tabor',
-      policy_id: '234.75.43.01',
-      referral_code: '324234 - AC',
-      referral_provider: 'Creek Hospital',
-      destination_provider: 'Creek Hospital',
-      status: 'Approved',
-      reason_for_request: 'Lorem ipsum dolor',
-    },
-    {
-      date: '27/10/21',
-      patients_name: 'Tejiri Tabor',
-      policy_id: '234.75.43.01',
-      referral_code: '324234 - AC',
-      referral_provider: 'Creek Hospital',
-      destination_provider: 'Creek Hospital',
-      status: 'Approved',
-      reason_for_request: 'Lorem ipsum dolor',
-    },
-    {
-      date: '27/10/21',
-      patients_name: 'Tejiri Tabor',
-      policy_id: '234.75.43.01',
-      referral_code: '324234 - AC',
-      referral_provider: 'Creek Hospital',
-      destination_provider: 'Creek Hospital',
-      status: 'Approved',
-      reason_for_request: 'Lorem ipsum dolor',
-    },
-  ];
+//     {
+//       date: '27/10/21',
+//       patients_name: 'Tejiri Tabor',
+//       policy_id: '234.75.43.01',
+//       referral_code: '324234 - AC',
+//       referral_provider: 'Creek Hospital',
+//       destination_provider: 'Creek Hospital',
+//       status: 'Approved',
+//       reason_for_request: 'Lorem ipsum dolor',
+//     },
+//     {
+//       date: '27/10/21',
+//       patients_name: 'Tejiri Tabor',
+//       policy_id: '234.75.43.01',
+//       referral_code: '324234 - AC',
+//       referral_provider: 'Creek Hospital',
+//       destination_provider: 'Creek Hospital',
+//       status: 'Approved',
+//       reason_for_request: 'Lorem ipsum dolor',
+//     },
+//     {
+//       date: '27/10/21',
+//       patients_name: 'Tejiri Tabor',
+//       policy_id: '234.75.43.01',
+//       referral_code: '324234 - AC',
+//       referral_provider: 'Creek Hospital',
+//       destination_provider: 'Creek Hospital',
+//       status: 'Approved',
+//       reason_for_request: 'Lorem ipsum dolor',
+//     },
+//     {
+//       date: '27/10/21',
+//       patients_name: 'Tejiri Tabor',
+//       policy_id: '234.75.43.01',
+//       referral_code: '324234 - AC',
+//       referral_provider: 'Creek Hospital',
+//       destination_provider: 'Creek Hospital',
+//       status: 'Approved',
+//       reason_for_request: 'Lorem ipsum dolor',
+//     },
+//     {
+//       date: '27/10/21',
+//       patients_name: 'Tejiri Tabor',
+//       policy_id: '234.75.43.01',
+//       referral_code: '324234 - AC',
+//       referral_provider: 'Creek Hospital',
+//       destination_provider: 'Creek Hospital',
+//       status: 'Approved',
+//       reason_for_request: 'Lorem ipsum dolor',
+//     },
+//   ];
 
-  const returnCell = (status) => {
-    // if (status === "approved") {
-    //   return <span style={{color: "green"}}>{status}</span>;
-    // }
-    // else if
-    switch (status.toLowerCase()) {
-      case 'active':
-        return <span style={{ color: '#17935C' }}>{status}</span>;
+//   const returnCell = (status) => {
+//     // if (status === "approved") {
+//     //   return <span style={{color: "green"}}>{status}</span>;
+//     // }
+//     // else if
+//     switch (status.toLowerCase()) {
+//       case 'active':
+//         return <span style={{ color: '#17935C' }}>{status}</span>;
 
-      case 'inactive':
-        return <span style={{ color: '#0364FF' }}>{status}</span>;
+//       case 'inactive':
+//         return <span style={{ color: '#0364FF' }}>{status}</span>;
 
-      default:
-        break;
-    }
-  };
+//       default:
+//         break;
+//     }
+//   };
 
-  const ReferralSchema = [
-    {
-      name: 'Date',
-      key: 'date',
-      description: 'Enter date',
-      selector: (row) => row.date,
-      sortable: true,
-      required: true,
-      inputType: 'DATE',
-    },
-    {
-      name: 'Patients Name',
-      key: 'patients_name',
-      description: 'Enter patients name',
-      selector: (row) => row.patients_name,
-      sortable: true,
-      required: true,
-      inputType: 'TEXT',
-    },
-    {
-      name: 'Policy ID',
-      key: 'policy_id',
-      description: 'Enter policy ID',
-      selector: (row) => row.policy_id,
-      sortable: true,
-      required: true,
-      inputType: 'TEXT',
-    },
-    {
-      name: 'Referral Code',
-      key: 'referral_code',
-      description: 'Enter referral code',
-      selector: (row) => row.referral_code,
-      sortable: true,
-      required: true,
-      inputType: 'TEXT',
-    },
-    {
-      name: 'Referral Provider',
-      key: 'referral_provider',
-      description: 'Enter referral provider',
-      selector: (row, i) => row.referral_provider,
-      sortable: true,
-      required: true,
-      inputType: 'DATE',
-    },
-    {
-      name: 'Destination Provider',
-      key: 'destination_provider',
-      description: 'Enter destination provider',
-      selector: (row, i) => row.destination_provider,
-      sortable: true,
-      required: true,
-      inputType: 'TEXT',
-    },
-    {
-      name: 'Status',
-      key: 'status',
-      description: 'Enter your status',
-      selector: (row, i) => row.status,
-      sortable: true,
-      required: true,
-      inputType: 'TEXT',
-    },
-    {
-      name: 'Reason for Request',
-      key: 'reason_for_request',
-      description: 'Enter the reason for the request',
-      selector: (row, i) => row.reason_for_request,
-      sortable: true,
-      required: true,
-      inputType: 'TEXT',
-    },
-  ];
-  const conditionalRowStyles = [
-    {
-      when: (row) => row.status === 'approved',
-      style: {
-        color: 'red',
-        '&:hover': {
-          cursor: 'pointer',
-        },
-      },
-    },
-    {
-      when: (row) => row.status === 'ongoing',
-      style: {
-        color: 'rgba(0,0,0,.54)',
-        '&:hover': {
-          cursor: 'pointer',
-        },
-      },
-    },
-    {
-      when: (row) => row.status === 'pending',
-      style: {
-        color: 'pink',
-        '&:hover': {
-          cursor: 'pointer',
-        },
-      },
-    },
-    {
-      when: (row) => row.status === 'declined',
-      style: {
-        color: 'purple',
-        backgroundColor: 'green',
-        '&:hover': {
-          cursor: 'pointer',
-        },
-      },
-    },
-  ];
+//   const ReferralSchema = [
+//     {
+//       name: 'Date',
+//       key: 'date',
+//       description: 'Enter date',
+//       selector: (row) => row.date,
+//       sortable: true,
+//       required: true,
+//       inputType: 'DATE',
+//     },
+//     {
+//       name: 'Patients Name',
+//       key: 'patients_name',
+//       description: 'Enter patients name',
+//       selector: (row) => row.patients_name,
+//       sortable: true,
+//       required: true,
+//       inputType: 'TEXT',
+//     },
+//     {
+//       name: 'Policy ID',
+//       key: 'policy_id',
+//       description: 'Enter policy ID',
+//       selector: (row) => row.policy_id,
+//       sortable: true,
+//       required: true,
+//       inputType: 'TEXT',
+//     },
+//     {
+//       name: 'Referral Code',
+//       key: 'referral_code',
+//       description: 'Enter referral code',
+//       selector: (row) => row.referral_code,
+//       sortable: true,
+//       required: true,
+//       inputType: 'TEXT',
+//     },
+//     {
+//       name: 'Referral Provider',
+//       key: 'referral_provider',
+//       description: 'Enter referral provider',
+//       selector: (row, i) => row.referral_provider,
+//       sortable: true,
+//       required: true,
+//       inputType: 'DATE',
+//     },
+//     {
+//       name: 'Destination Provider',
+//       key: 'destination_provider',
+//       description: 'Enter destination provider',
+//       selector: (row, i) => row.destination_provider,
+//       sortable: true,
+//       required: true,
+//       inputType: 'TEXT',
+//     },
+//     {
+//       name: 'Status',
+//       key: 'status',
+//       description: 'Enter your status',
+//       selector: (row, i) => row.status,
+//       sortable: true,
+//       required: true,
+//       inputType: 'TEXT',
+//     },
+//     {
+//       name: 'Reason for Request',
+//       key: 'reason_for_request',
+//       description: 'Enter the reason for the request',
+//       selector: (row, i) => row.reason_for_request,
+//       sortable: true,
+//       required: true,
+//       inputType: 'TEXT',
+//     },
+//   ];
+//   const conditionalRowStyles = [
+//     {
+//       when: (row) => row.status === 'approved',
+//       style: {
+//         color: 'red',
+//         '&:hover': {
+//           cursor: 'pointer',
+//         },
+//       },
+//     },
+//     {
+//       when: (row) => row.status === 'ongoing',
+//       style: {
+//         color: 'rgba(0,0,0,.54)',
+//         '&:hover': {
+//           cursor: 'pointer',
+//         },
+//       },
+//     },
+//     {
+//       when: (row) => row.status === 'pending',
+//       style: {
+//         color: 'pink',
+//         '&:hover': {
+//           cursor: 'pointer',
+//         },
+//       },
+//     },
+//     {
+//       when: (row) => row.status === 'declined',
+//       style: {
+//         color: 'purple',
+//         backgroundColor: 'green',
+//         '&:hover': {
+//           cursor: 'pointer',
+//         },
+//       },
+//     },
+//   ];
 
-  return (
-    <>
-      {user ? (
-        <>
-          <div className="level">
-            <PageWrapper
-              style={{ flexDirection: 'column', padding: '0.6rem 1rem' }}
-            >
-              <TableMenu>
-                <div style={{ display: 'flex', alignItems: 'center' }}>
-                  {handleSearch && (
-                    <div className="inner-table">
-                      <FilterMenu onSearch={handleSearch} />
-                    </div>
-                  )}
-                  <h2 style={{ margin: '0 10px', fontSize: '0.95rem' }}>
-                    Referral
-                  </h2>
-                  <DatePicker
-                    selected={startDate}
-                    onChange={(date) => handleDate(date)}
-                    dateFormat="dd/MM/yyyy"
-                    placeholderText="Filter By Date"
-                    isClearable
-                  />
-                  {/* <SwitchButton /> */}
-                  <Switch>
-                    <button
-                      value={value}
-                      onClick={() => {
-                        setValue('list');
-                      }}
-                      style={value === 'list' ? activeStyle : {}}
-                    >
-                      <BsList style={{ fontSize: '1rem' }} />
-                    </button>
-                    <button
-                      value={value}
-                      onClick={() => {
-                        setValue('grid');
-                      }}
-                      style={value === 'grid' ? activeStyle : {}}
-                    >
-                      <BsFillGridFill style={{ fontSize: '1rem' }} />
-                    </button>
-                  </Switch>
-                </div>
+//   return (
+//     <>
+//       {user ? (
+//         <>
+//           <div className="level">
+//             <PageWrapper
+//               style={{ flexDirection: 'column', padding: '0.6rem 1rem' }}
+//             >
+//               <TableMenu>
+//                 <div style={{ display: 'flex', alignItems: 'center' }}>
+//                   {handleSearch && (
+//                     <div className="inner-table">
+//                       <FilterMenu onSearch={handleSearch} />
+//                     </div>
+//                   )}
+//                   <h2 style={{ margin: '0 10px', fontSize: '0.95rem' }}>
+//                     Referral
+//                   </h2>
+//                   <DatePicker
+//                     selected={startDate}
+//                     onChange={(date) => handleDate(date)}
+//                     dateFormat="dd/MM/yyyy"
+//                     placeholderText="Filter By Date"
+//                     isClearable
+//                   />
+//                   {/* <SwitchButton /> */}
+//                   <Switch>
+//                     <button
+//                       value={value}
+//                       onClick={() => {
+//                         setValue('list');
+//                       }}
+//                       style={value === 'list' ? activeStyle : {}}
+//                     >
+//                       <BsList style={{ fontSize: '1rem' }} />
+//                     </button>
+//                     <button
+//                       value={value}
+//                       onClick={() => {
+//                         setValue('grid');
+//                       }}
+//                       style={value === 'grid' ? activeStyle : {}}
+//                     >
+//                       <BsFillGridFill style={{ fontSize: '1rem' }} />
+//                     </button>
+//                   </Switch>
+//                 </div>
 
-                {handleCreateNew && (
-                  <GlobalCustomButton onClick={handleCreateNew}>
-                    <AddCircleOutline
-                      sx={{ marginRight: '5px' }}
-                      fontSize="small"
-                    />
-                    Add New Referral
-                  </GlobalCustomButton>
-                )}
-              </TableMenu>
-              {value === 'list' ? (
-                <CustomTable
-                  title={''}
-                  columns={ReferralSchema}
-                  data={dummyData}
-                  pointerOnHover
-                  highlightOnHover
-                  striped
-                  onRowClicked={handleRow}
-                  progressPending={loading}
-                  //conditionalRowStyles={conditionalRowStyles}
-                />
-              ) : (
-                <CalendarGrid appointments={mapFacilities()} />
-              )}
-            </PageWrapper>
-          </div>
-        </>
-      ) : (
-        <div>loading</div>
-      )}
-    </>
-  );
-}
+//                 {handleCreateNew && (
+//                   <GlobalCustomButton onClick={handleCreateNew}>
+//                     <AddCircleOutline
+//                       sx={{ marginRight: '5px' }}
+//                       fontSize="small"
+//                     />
+//                     Add New Referral
+//                   </GlobalCustomButton>
+//                 )}
+//               </TableMenu>
+//               {value === 'list' ? (
+//                 <CustomTable
+//                   title={''}
+//                   columns={ReferralSchema}
+//                   data={dummyData}
+//                   pointerOnHover
+//                   highlightOnHover
+//                   striped
+//                   onRowClicked={handleRow}
+//                   progressPending={loading}
+//                   //conditionalRowStyles={conditionalRowStyles}
+//                 />
+//               ) : (
+//                 <CalendarGrid appointments={mapFacilities()} />
+//               )}
+//             </PageWrapper>
+//           </div>
+//         </>
+//       ) : (
+//         <div>loading</div>
+//       )}
+//     </>
+//   );
+// }
+
 export function Details({ showModal, setShowModal }) {
   const [deny, setDeny] = useState(false);
   const [approve, setApprove] = useState(false);
@@ -1198,46 +1215,46 @@ export function Details({ showModal, setShowModal }) {
     <>
       <div
         style={{
-          width: '100%',
-          height: 'calc(100vh - 90px)',
-          overflow: 'auto',
-          paddingRight: '1rem',
+          width: "100%",
+          height: "calc(100vh - 90px)",
+          overflow: "auto",
+          paddingRight: "1rem",
         }}
       >
         <Box
           sx={{
-            display: 'flex',
-            justifyContent: 'space-between',
-            alignItems: 'center',
+            display: "flex",
+            justifyContent: "space-between",
+            alignItems: "center",
           }}
         >
           <Box>
-            <FormsHeaderText text={'Referral Details - 13322BA'} />
+            <FormsHeaderText text={"Referral Details - 13322BA"} />
           </Box>
-          <Box sx={{ display: 'flex', marginTop: '1rem' }}>
+          <Box sx={{ display: "flex", marginTop: "1rem" }}>
             <GlobalCustomButton
               text="Back"
               onClick={() => setShowModal(0)}
               color="warning"
-              customStyles={{ marginRight: '.8rem' }}
+              customStyles={{ marginRight: ".8rem" }}
             />
             <GlobalCustomButton
               onClick={() => setApprove(true)}
               text="Approve"
               color="success"
-              customStyles={{ marginRight: '.8rem' }}
+              customStyles={{ marginRight: ".8rem" }}
             />
             <GlobalCustomButton
               onClick={() => {}}
               text="On Hold"
               color="secondary"
-              customStyles={{ marginRight: '.8rem' }}
+              customStyles={{ marginRight: ".8rem" }}
             />
             <GlobalCustomButton
               onClick={() => setDeny(true)}
               text="Reject"
               color="error"
-              customStyles={{ marginRight: '.8rem' }}
+              customStyles={{ marginRight: ".8rem" }}
             />
             <GlobalCustomButton
               onClick={
@@ -1245,14 +1262,14 @@ export function Details({ showModal, setShowModal }) {
                   ? () => setCurrentPage(2)
                   : () => setCurrentPage(1)
               }
-              text={currentPage === 1 ? 'Task' : 'Details'}
+              text={currentPage === 1 ? "Task" : "Details"}
               variant="outlined"
-              customStyles={{ marginRight: '.8rem' }}
+              customStyles={{ marginRight: ".8rem" }}
             />
             <Badge
               badgeContent={4}
               color="success"
-              sx={{ marginRight: '10px' }}
+              sx={{ marginRight: "10px" }}
             >
               <GlobalCustomButton
                 onClick={() => setOpenDrawer(true)}
@@ -1265,9 +1282,9 @@ export function Details({ showModal, setShowModal }) {
         {currentPage === 1 && (
           <div
             style={{
-              marginTop: '10px',
-              border: '1px solid #8F8F8F',
-              padding: '1rem',
+              marginTop: "10px",
+              border: "1px solid #8F8F8F",
+              padding: "1rem",
             }}
           >
             <p>Request Sent 08/05/2022 9:45pm</p>
@@ -1297,11 +1314,11 @@ export function Details({ showModal, setShowModal }) {
                 <p>Fee for Service: Applicable</p>
               </Grid>
             </Grid>
-            <FormsHeaderText text={'Referral Code - 13322BA'} />
-            <McText txt={'Clinical Information'} />
+            <FormsHeaderText text={"Referral Code - 13322BA"} />
+            <McText txt={"Clinical Information"} />
             <Grid container spacing={2} mb={1}>
               <Grid item xs={12}>
-                <p style={{ fontWeight: 'bold' }}>Presenting Complaints:</p>
+                <p style={{ fontWeight: "bold" }}>Presenting Complaints:</p>
                 <p>
                   Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed
                   do eiusmod tempor incididunt
@@ -1309,23 +1326,23 @@ export function Details({ showModal, setShowModal }) {
               </Grid>
             </Grid>
 
-            <FormsHeaderText text={'Clinical Findings'} />
+            <FormsHeaderText text={"Clinical Findings"} />
             <Grid container spacing={2} mb={1}>
               <Grid item xs={12}>
-                <p style={{ fontWeight: 'bold' }}>Provisional Diagonosis:</p>
+                <p style={{ fontWeight: "bold" }}>Provisional Diagonosis:</p>
                 <p>
                   Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed
                   do eiusmod tempor incididunt
                 </p>
 
-                <p style={{ fontWeight: 'bold' }}>
+                <p style={{ fontWeight: "bold" }}>
                   Planned Procedures / Services Requiring Authorization:
                 </p>
                 <p>
                   Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed
                   do eiusmod tempor incididunt
                 </p>
-                <p style={{ fontWeight: 'bold' }}>
+                <p style={{ fontWeight: "bold" }}>
                   Planned Procedures / Services Requiring Authorization:
                 </p>
               </Grid>
@@ -1333,24 +1350,24 @@ export function Details({ showModal, setShowModal }) {
 
             <Grid container spacing={2}>
               <Grid item xs={12}>
-                <p style={{ fontWeight: 'bold' }}>Reason for Request:</p>
+                <p style={{ fontWeight: "bold" }}>Reason for Request:</p>
                 <span
                   style={{
-                    fontWeight: 'bold',
-                    backgroundColor: '#ECF3FF',
-                    color: '#0364FF',
-                    padding: '.3rem',
-                    marginRight: '1rem',
+                    fontWeight: "bold",
+                    backgroundColor: "#ECF3FF",
+                    color: "#0364FF",
+                    padding: ".3rem",
+                    marginRight: "1rem",
                   }}
                 >
                   Procedure
                 </span>
                 <span
                   style={{
-                    fontWeight: 'bold',
-                    backgroundColor: '#ECF3FF',
-                    color: '#0364FF',
-                    padding: '.3rem',
+                    fontWeight: "bold",
+                    backgroundColor: "#ECF3FF",
+                    color: "#0364FF",
+                    padding: ".3rem",
                   }}
                 >
                   Services
@@ -1359,7 +1376,7 @@ export function Details({ showModal, setShowModal }) {
             </Grid>
             <Grid container spacing={2}>
               <Grid item xs={12}>
-                <p style={{ fontWeight: 'bold' }}>Physician Name:</p>
+                <p style={{ fontWeight: "bold" }}>Physician Name:</p>
                 <p>Dr. John Doe</p>
                 <p>Lagos State Hospital</p>
               </Grid>
@@ -1367,7 +1384,7 @@ export function Details({ showModal, setShowModal }) {
           </div>
         )}
         {currentPage === 2 && (
-          <div style={{ marginTop: '1rem' }}>
+          <div style={{ marginTop: "1rem" }}>
             <CRMTasks />
           </div>
         )}
@@ -1379,10 +1396,10 @@ export function Details({ showModal, setShowModal }) {
               <ModalHeader text={`Approve Claim  13229-BA`} />
               <Grid container spacing={2} mt={1}>
                 <Grid item xs={12}>
-                  <Textarea label={'Comment'} />
+                  <Textarea label={"Comment"} />
                 </Grid>
                 <Grid item xs={12}>
-                  <GlobalCustomButton text={'Approve'} color="success" />
+                  <GlobalCustomButton text={"Approve"} color="success" />
                 </Grid>
               </Grid>
             </form>
@@ -1397,10 +1414,10 @@ export function Details({ showModal, setShowModal }) {
 
               <Grid container spacing={2} mt={1}>
                 <Grid item xs={12}>
-                  <Textarea label={'Comment'} />
+                  <Textarea label={"Comment"} />
                 </Grid>
                 <Grid item xs={12}>
-                  <GlobalCustomButton text={'Reject'} color="error" />
+                  <GlobalCustomButton text={"Reject"} color="error" />
                 </Grid>
               </Grid>
             </form>
@@ -1410,11 +1427,11 @@ export function Details({ showModal, setShowModal }) {
       <Drawer
         open={openDrawer}
         sx={{
-          width: 'fit-content',
+          width: "fit-content",
           // height: 'fit-content',
           flexShrink: 0,
-          '& .MuiDrawer-paper': {
-            width: 'fit-content',
+          "& .MuiDrawer-paper": {
+            width: "fit-content",
             // height: 'fit-content',
           },
         }}
@@ -1423,9 +1440,9 @@ export function Details({ showModal, setShowModal }) {
       >
         <Box
           sx={{
-            width: '25vw',
-            height: '100vh',
-            overflowY: 'hidden',
+            width: "25vw",
+            height: "100vh",
+            overflowY: "hidden",
           }}
         >
           <ChatInterface closeChat={() => setOpenDrawer(false)} />
@@ -1442,14 +1459,14 @@ export function ReferralModify({ showModal, setShowModal }) {
   const [success, setSuccess] = useState(false);
   const [success1, setSuccess1] = useState(false);
   const [success2, setSuccess2] = useState(false);
-  const [message, setMessage] = useState('');
+  const [message, setMessage] = useState("");
   const [clientId, setClientId] = useState();
   const [locationId, setLocationId] = useState();
   const [practionerId, setPractionerId] = useState();
   const [type, setType] = useState();
   // eslint-disable-next-line
   const [facility, setFacility] = useState();
-  const ClientServ = client.service('appointments');
+  const ClientServ = client.service("appointments");
   //const navigate=useNavigate()
   const { user } = useContext(UserContext); //,setUser
   // eslint-disable-next-line
@@ -1457,14 +1474,14 @@ export function ReferralModify({ showModal, setShowModal }) {
   const [selectedClient, setSelectedClient] = useState();
   const [selectedAppointment, setSelectedAppointment] = useState();
   // const [appointment_reason,setAppointment_reason]= useState()
-  const [appointment_status, setAppointment_status] = useState('');
-  const [appointment_type, setAppointment_type] = useState('');
+  const [appointment_status, setAppointment_status] = useState("");
+  const [appointment_type, setAppointment_type] = useState("");
   const [billingModal, setBillingModal] = useState(false);
 
   const [chosen, setChosen] = useState();
   const [chosen1, setChosen1] = useState();
   const [chosen2, setChosen2] = useState();
-  const appClass = ['On-site', 'Teleconsultation', 'Home Visit'];
+  const appClass = ["On-site", "Teleconsultation", "Home Visit"];
 
   let appointee; //  =state.ClientModule.selectedClient
   /*  const getSearchfacility=(obj)=>{
@@ -1537,7 +1554,7 @@ export function ReferralModify({ showModal, setShowModal }) {
 
   const onSubmit = (data, e) => {
     e.preventDefault();
-    setMessage('');
+    setMessage("");
     setError(false);
     setSuccess(false);
     setShowModal(false),
@@ -1545,7 +1562,7 @@ export function ReferralModify({ showModal, setShowModal }) {
         ...prevstate,
         AppointmentModule: {
           selectedAppointment: {},
-          show: 'list',
+          show: "list",
         },
       }));
 
@@ -1567,7 +1584,7 @@ export function ReferralModify({ showModal, setShowModal }) {
     data.gender = chosen.gender;
     data.phone = chosen.phone;
     data.email = chosen.email;
-    data.practitioner_name = chosen2.firstname + ' ' + chosen2.lastname;
+    data.practitioner_name = chosen2.firstname + " " + chosen2.lastname;
     data.practitioner_profession = chosen2.profession;
     data.practitioner_department = chosen2.department;
     data.location_name = chosen1.name;
@@ -1584,18 +1601,18 @@ export function ReferralModify({ showModal, setShowModal }) {
       .then((res) => {
         //console.log(JSON.stringify(res))
         e.target.reset();
-        setAppointment_type('');
-        setAppointment_status('');
-        setClientId('');
-        setLocationId('');
+        setAppointment_type("");
+        setAppointment_status("");
+        setClientId("");
+        setLocationId("");
         /*  setMessage("Created Client successfully") */
         setSuccess(true);
         setSuccess1(true);
         setSuccess2(true);
         toast({
           message:
-            'Appointment created succesfully, Kindly bill patient if required',
-          type: 'is-success',
+            "Appointment created succesfully, Kindly bill patient if required",
+          type: "is-success",
           dismissible: true,
           pauseOnHover: true,
         });
@@ -1606,8 +1623,8 @@ export function ReferralModify({ showModal, setShowModal }) {
       })
       .catch((err) => {
         toast({
-          message: 'Error creating Appointment ' + err,
-          type: 'is-danger',
+          message: "Error creating Appointment " + err,
+          type: "is-danger",
           dismissible: true,
           pauseOnHover: true,
         });
@@ -1641,8 +1658,8 @@ export function ReferralModify({ showModal, setShowModal }) {
             } */
   const CustomSelectData = [
     {
-      label: 'Today',
-      value: 'today',
+      label: "Today",
+      value: "today",
     },
   ];
 
@@ -1650,12 +1667,12 @@ export function ReferralModify({ showModal, setShowModal }) {
     <>
       <div
         className="card "
-        style={{ height: '100%', width: '80vw', overflowX: 'hidden' }}
+        style={{ height: "100%", width: "80vw", overflowX: "hidden" }}
       >
         <form onSubmit={handleSubmit(onSubmit)}>
           <Grid container spacing={2}>
             <Grid item xs={12} sm={6}>
-              <ModalHeader text={'Referral Modify'} />
+              <ModalHeader text={"Referral Modify"} />
             </Grid>
             <Grid item xs={12} sm={6}>
               <MdCancel
@@ -1665,15 +1682,15 @@ export function ReferralModify({ showModal, setShowModal }) {
                       ...prevstate,
                       AppointmentModule: {
                         selectedAppointment: {},
-                        show: 'list',
+                        show: "list",
                       },
                     }));
                 }}
                 style={{
-                  fontSize: '2rem',
-                  color: 'crimson',
-                  cursor: 'pointer',
-                  float: 'right',
+                  fontSize: "2rem",
+                  color: "crimson",
+                  cursor: "pointer",
+                  float: "right",
                 }}
               />
             </Grid>
@@ -1691,10 +1708,10 @@ export function ReferralModify({ showModal, setShowModal }) {
 
           <div
             style={{
-              display: 'flex',
-              alignItems: 'center',
-              gap: '5rem',
-              marginBottom: '0.6rem',
+              display: "flex",
+              alignItems: "center",
+              gap: "5rem",
+              marginBottom: "0.6rem",
             }}
           >
             <BasicDateTimePicker name="dateOfRequest" label="Date of Request" />
@@ -1705,9 +1722,9 @@ export function ReferralModify({ showModal, setShowModal }) {
                 defaultValue="emergency"
                 name="emergency"
                 sx={{
-                  display: 'flex !important',
-                  justifyContent: 'space-between !',
-                  flexDirection: 'row !important',
+                  display: "flex !important",
+                  justifyContent: "space-between !",
+                  flexDirection: "row !important",
                 }}
               >
                 <FormControlLabel value="yes" control={<Radio />} label="Yes" />
@@ -1718,27 +1735,27 @@ export function ReferralModify({ showModal, setShowModal }) {
 
           <div>
             <McText
-              txt={'Clinical Information'}
-              color={'#0064CC'}
-              type={'p'}
-              bold={'700'}
-              size={'18px'}
+              txt={"Clinical Information"}
+              color={"#0064CC"}
+              type={"p"}
+              bold={"700"}
+              size={"18px"}
             />
           </div>
           <GridBox>
             <button
               style={{
-                float: 'right',
-                backgroundColor: '#ECF3FF',
-                color: '#0064CC',
-                border: 'none',
-                padding: '10px',
-                cursor: 'pointer',
+                float: "right",
+                backgroundColor: "#ECF3FF",
+                color: "#0064CC",
+                border: "none",
+                padding: "10px",
+                cursor: "pointer",
               }}
             >
               <MdAddCircle
                 style={{
-                  marginRight: '5px',
+                  marginRight: "5px",
                 }}
               />
               Add complaints
@@ -1759,11 +1776,11 @@ export function ReferralModify({ showModal, setShowModal }) {
           <Grid container spacing={2} my={2}>
             <Grid item xs={12} sm={12}>
               <McText
-                txt={'Clinic Findings'}
-                color={'#0064CC'}
-                type={'p'}
-                bold={'700'}
-                size={'18px'}
+                txt={"Clinic Findings"}
+                color={"#0064CC"}
+                type={"p"}
+                bold={"700"}
+                size={"18px"}
               />
             </Grid>
           </Grid>
@@ -1777,17 +1794,17 @@ export function ReferralModify({ showModal, setShowModal }) {
 
             <button
               style={{
-                float: 'left',
-                backgroundColor: '#ECF3FF',
-                color: '#0064CC',
-                border: 'none',
-                padding: '10px',
-                cursor: 'pointer',
+                float: "left",
+                backgroundColor: "#ECF3FF",
+                color: "#0064CC",
+                border: "none",
+                padding: "10px",
+                cursor: "pointer",
               }}
             >
               <MdAddCircle
                 style={{
-                  marginRight: '5px',
+                  marginRight: "5px",
                 }}
               />
               Add Diagnosis
@@ -1801,17 +1818,17 @@ export function ReferralModify({ showModal, setShowModal }) {
 
             <button
               style={{
-                float: 'left',
-                backgroundColor: '#ECF3FF',
-                color: '#0064CC',
-                border: 'none',
-                padding: '10px',
-                cursor: 'pointer',
+                float: "left",
+                backgroundColor: "#ECF3FF",
+                color: "#0064CC",
+                border: "none",
+                padding: "10px",
+                cursor: "pointer",
               }}
             >
               <MdAddCircle
                 style={{
-                  marginRight: '5px',
+                  marginRight: "5px",
                 }}
               />
               Add Procedure
@@ -1825,17 +1842,17 @@ export function ReferralModify({ showModal, setShowModal }) {
 
             <button
               style={{
-                float: 'left',
-                backgroundColor: '#ECF3FF',
-                color: '#0064CC',
-                border: 'none',
-                padding: '10px',
-                cursor: 'pointer',
+                float: "left",
+                backgroundColor: "#ECF3FF",
+                color: "#0064CC",
+                border: "none",
+                padding: "10px",
+                cursor: "pointer",
               }}
             >
               <MdAddCircle
                 style={{
-                  marginRight: '5px',
+                  marginRight: "5px",
                 }}
               />
               Add Service
@@ -1843,10 +1860,10 @@ export function ReferralModify({ showModal, setShowModal }) {
           </GridBox>
           <div
             style={{
-              display: 'flex',
-              flexDirection: 'column',
-              gap: '0.6rem',
-              marginBottom: '10px',
+              display: "flex",
+              flexDirection: "column",
+              gap: "0.6rem",
+              marginBottom: "10px",
             }}
           >
             <Input name="physicianName" label="Physician' Name" type="text" />
